+++ src/components/Badge/types.ts
@@ -2,7 +2,7 @@ import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { UserInterfaceType } from 'types';
 
 export interface BadgeUIType extends UserInterfaceType {
-  badgeValue?: string;
+  badgeValue?: string | React.ReactNode;
   badgeClassName?: string;
   badgeIcon?: IconProp;
 }