+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.0.1-beta.4]](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/10)] - 2024-01-05
+
+- [Show fixed value instead of Scientific Notation, Display Fixes](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/9)
+
 ## [[0.0.1-beta.3]](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/8)] - 2023-12-05
 
 - [Added some more testids to assist testing](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/7)